move int conversion

+++ gym/envs/box2d/car_racing.py
@@ -369,9 +369,9 @@ class CarRacing(gym.Env, EzPickle):
         else:
             pixel_scale = 1
             if hasattr(win.context, '_nscontext'):
-                pixel_scale = int(win.context._nscontext.view().backingScaleFactor())  # pylint: disable=protected-access
-            VP_W = pixel_scale * WINDOW_W
-            VP_H = pixel_scale * WINDOW_H
+                pixel_scale = win.context._nscontext.view().backingScaleFactor()  # pylint: disable=protected-access
+            VP_W = int(pixel_scale * WINDOW_W)
+            VP_H = int(pixel_scale * WINDOW_H)
 
         gl.glViewport(0, 0, VP_W, VP_H)
         t.enable()
